Boon Prefab and Currency
Added boon prefab and boons prefab. Boon can be placed anywhere and boons are a row of three for the end of the level. Need to add display to show what each on does and change pick up from trigger to interaction. NOTE: Boons should be placed in a parent called Boons so that it can get destroyed and as well as all other children boons.

Currency added to playerstats. Whenever a player picks up a coin, they are rewarded a flat 5 coins as currency.